Update config.md
I think we need to rephrase a bit the definitions of the min/max power dynamics. The way I understood it is that max is the maximum (in percentage) allowed power for discharge. While min is still the maximum but for discharge. Otherwise max should be the maximum allowed variation wile min the minimum allowed variation (applicable both for charge and discharge), but this is not the way I understood it from your explanation in a post of the thread